Consolidate domain-merge review into one progressive-disclosure surface
Two review surfaces had diverged: the older /admin/domains (flat five pills:
Keep A / Keep B / Leave / A under B / B under A) and the newer
/admin/knowledge/merge-review (progressive disclosure, Dismiss + undo, but
graph-node-only). This folds the best of both into a single surface at
/admin/domains and removes the duplicate.

The consolidated /admin/domains now has:
- Progressive disclosure: a collapsed row shows only Merge/Nest + a quiet
  Dismiss; expanding reveals the four choices grouped under Same-scope (merge)
  / Different-scope (nest) radios, with Preview inline. One row expands at a
  time; the ★ keep marker and in-graph / label-only chips stay.
- The raw-corpus-label backend (/api/admin/domain-merges): merge folds labels
  with no graph node via the census-guarded shared applier, and Nest authors
  both sides into the graph (creating nodes as needed) — capabilities the
  graph-only version lacked.
- Dismiss writes the permanent reviewed-pair record (/api/admin/knowledge
  dismiss_pair) so a genuinely-different pair never resurfaces.
- A persist-with-undo toast: merge/nest apply optimistically via a deferred
  client-side commit — the API call fires only after the ~6s window elapses,
  so Undo restores the row with nothing committed (the only honest undo for an
  irreversible corpus merge). One toast at a time; a new action auto-commits
  the prior.

Removed the duplicate /admin/knowledge/merge-review page + client and the link
into it; updated the /admin/domains intro copy for the Dismiss/undo model.

Typecheck, lint, and the font/color/spacing/radius ratchets pass; the
order-independent reviewedPairKey unit test still passes.
